fix: changed file names by moving RandomWords.py contents into WordFinder.py
